Cross-cutting cleanup: extract fit-chip, drop dead code, drop tombstone tests
- fit-chip.ts is the single home for the fits/tight/won't-run chip; model-card
  and model-detail both call renderFitChip(container, entry.fit) instead of
  duplicating the guard and label switch.
- OCR_BACKEND const and OcrBackend type were declared but unused. Deleted.
- MODEL_INFO_NO_ACTIVE locale string was unused; only NOTICE_NO_ACTIVE_MODEL is
  read. Deleted.
- ModelsResponse 'older servers' comment was iteration history. Dropped.
- Parity tests asserting the absence of removed syncMode/SYNC_MODE were
  tombstones. The type system already enforces removal. Deleted.
